+++ tests/resources/database/migrations/0000_00_00_000006_transactions.php
@@ -21,12 +21,12 @@ class Transactions extends Migration
     public function up()
     {
         Schema::create('transactions', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('billable_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('billable_id');
             $table->foreign('billable_id')->references('id')->on($this->billableTableName);
-            $table->unsignedInteger('credit_card_id');
+            $table->unsignedBigInteger('credit_card_id');
             $table->foreign('credit_card_id')->references('id')->on('credit_cards');
-            $table->unsignedInteger('subscription_id')->nullable();
+            $table->unsignedBigInteger('subscription_id')->nullable();
             $table->foreign('subscription_id')->references('id')->on('subscriptions');
             $table->double('amount');
             $table->string('currency', 3)->default('TRY');
